+++ components/Login/SignIn.tsx
@@ -2,14 +2,7 @@ import React from 'react'
 import { Form, Formik, Field } from 'formik'
 import { StyledTextInput } from '../FormStyles/InputBox'
 import * as Yup from 'yup'
-import {
-  BackButton,
-  EStyledButton,
-  Container,
-  SubContainer,
-  Title,
-  SubTitle,
-} from './LoginStyling'
+import { EStyledButton, Container, SubContainer } from './LoginStyling'
 import styled from 'styled-components'
 import Link from 'next/link'
 import { PasswordInputBox } from '../FormStyles/PasswordInputBox'
@@ -29,7 +22,6 @@ export const RememberDiv = styled.div`
 interface FormValues {
   email: string
   password: string
-  checked: Array<string>
 }
 
 // component for the signin page - includes form (and styling) for validation
@@ -37,7 +29,6 @@ function SignIn() {
   const initialVal: FormValues = {
     email: '',
     password: '',
-    checked: [],
   }
 
   return (
@@ -56,43 +47,25 @@ function SignIn() {
     >
       <Form>
         <Container>
-          <Link href="/home" passHref>
-            <BackButton type="button">&lt; Back</BackButton>
-          </Link>
-          <Title>Welcome back!</Title>
-          <SubTitle>Please sign in.</SubTitle>
           <SubContainer>
             <StyledTextInput
-              width={600}
-              height={53}
+              width={356}
+              height={42}
               name="email"
               placeholder="Email"
               type="text"
             />
 
             <PasswordInputBox
-              width={600}
-              height={53}
+              width={356}
+              height={42}
               placeholder="Password"
               name="password"
             />
           </SubContainer>
-          <Link href="/forgotpassword" passHref>
-            <p>
-              Forgot <strong>Password?</strong>
-            </p>
-          </Link>
-          <RememberSignIn>
-            <RememberDiv>
-              <label>
-                <Field type="checkbox" name="checked" value="Remember" />
-                Remember me
-              </label>
-            </RememberDiv>
-            <EStyledButton type="submit" primary={true}>
-              Sign In
-            </EStyledButton>
-          </RememberSignIn>
+          <EStyledButton type="submit" primary={true}>
+            Sign In
+          </EStyledButton>
         </Container>
       </Form>
     </Formik>
