+++ pages/signup.tsx
@@ -1,24 +1,12 @@
 import SignUp from '../components/Login/SignUp'
 import LoginContainer from '../components/Login/LoginContainer'
 import React from 'react'
-import styled from 'styled-components'
-import { COLORS } from '../constants/colors'
-import NavBar from '../components/Critical/NavBar'
-
-const LoginPageDiv = styled.div`
-  width: 100%;
-  height: calc(100vh - 80px);
-  display: flex;
-  justify-content: center;
-  padding-top: 50px;
-  background-color: ${COLORS.LIGHT_GREY};
-`
+import LoginAbstraction from '../components/Login/LoginAbstraction'
 
 function signup() {
   return (
-    <div>
-      <NavBar />
-      <LoginPageDiv>
+    <LoginAbstraction
+      main={
         <LoginContainer
           headerPhrase="Create your account"
           bottomPhrase="Let's get you logged in"
@@ -26,8 +14,8 @@ function signup() {
           buttonLink="/signin"
           form={<SignUp />}
         />
-      </LoginPageDiv>
-    </div>
+      }
+    />
   )
 }
